refactor(hooks): move getUserHooksFor into the hooks registry
User-land hook lookup lived in plugins/mtime-cache.ts even though
discovery is the only cache concern. Move getUserHookEntriesFor and
getUserHooksFor next to getHooksFor, and expose discovered plugin
names from the cache so hook resolution stays in the hooks package.

No behavior change.
